agents/search-service/requirements.txt
# Search Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Azure SDK
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Monitoring
azure-monitor-opentelemetry==1.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
